+++ src/modules/FixTextObjectBounds.cpp
@@ -6,6 +6,7 @@ void FTOBTextGameObject::fixBounds() {
         auto ref = static_cast<RITextGameObject*>(static_cast<TextGameObject*>(this));
         if (ref->isReferenceImage()) return;
     }
+    
     m_width = getContentWidth();
     m_height = getContentHeight();
     updateOrientedBox();