+++ changelog.md
@@ -1,5 +1,10 @@
+# 1.0.0-beta.4
+- Fix a crash
+- Change Single Deselect to use Keybind
+- Fix Reference Image culling
+
 # 1.0.0-beta.3
-- Fix search input memory leak
+- Fix search input memory leak and crash
 
 # 1.0.0-beta.2
 - Fix a few Reference Image bugs